Completed second core quest

+++ Assets/[Scripts]/AbilityCooldowns.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 
 public class AbilityCooldowns : MonoBehaviour
 {
-    public void OriginalSpeed(PlayerController player, int speed)
+    public void OriginalSpeed(PlayerController player, float speed)
     {
         StartCoroutine(Speed(player, speed));
     }
 
-    IEnumerator Speed(PlayerController player, int speed)
+    IEnumerator Speed(PlayerController player, float speed)
     {
         yield return new WaitForSeconds(5);
         player.stats.MoveSpeed -= speed;
